calarieapp/requirements_minimal.txt
# Minimal requirements for AI Calorie Tracker (TensorFlow-free version)
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI and ML libraries (TensorFlow-free)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# LangChain and LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1

# Computer Vision
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Data processing and visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# UI and utilities
streamlit-option-menu>=0.3.0
gradio>=3.40.0
accelerate>=0.20.0
peft>=0.4.0

# Additional utilities
ipykernel>=6.25.0
validators==0.28.1
pytube>=15.0.0
